complete inventory management for purchases

+++ backend/app/models/cash_transactions.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
@@ -23,7 +24,7 @@ class CashTransaction(Base):
         nullable=True
     )
 
-    amount: Mapped[float] = mapped_column(
+    amount: Mapped[Decimal] = mapped_column(
         Numeric(12, 2)
     )
